app, 홈페이지 분리
